prototype-videolog | move video above header area
First, I took out the cancellation (remove_action) that prevents normal genesis standard layouts from displaying)

Then I added an action to based on the genesis code to position itself above the title (on the header). The key takeaway here is to add the number (8 or lower) so it supercedes priority over the title.